refactor: unify duplicated TOON parsing grammar
InMemoryTOONDataSource and IterableTOONDataSource each reimplemented the
entire TOON grammar state machine (array headers, tabular rows, nested
arrays, nested objects, indentation/exit rules), sharing only token-level
helpers via ToonSyntax. The two forks could drift, so a fix in one would
not reach the other.

Extract the line-driven grammar into a shared ToonFlattener that pushes
flattened key/value pairs to a BiConsumer sink. The iterable source feeds
it one line at a time and drains the buffered pairs, keeping memory bounded;
the in-memory source feeds every line and collects into its map. Both now
drive the same grammar instead of duplicating it.

Behavior is unchanged: all 378 data-module tests pass, including the
cross-check asserting the in-memory and iterable flattenings are equal.
